move version diff to banner
