+++ web/src/user/UserInterface.ts
@@ -13,8 +13,8 @@ import "@goauthentik/components/ak-nav-buttons";
 import { AKElement } from "@goauthentik/elements/Base";
 import { AuthenticatedInterface } from "@goauthentik/elements/Interface";
 import "@goauthentik/elements/ak-locale-context";
+import "@goauthentik/elements/banner/EnterpriseStatusBanner";
 import "@goauthentik/elements/buttons/ActionButton";
-import "@goauthentik/elements/enterprise/EnterpriseStatusBanner";
 import "@goauthentik/elements/messages/MessageContainer";
 import "@goauthentik/elements/notifications/APIDrawer";
 import "@goauthentik/elements/notifications/NotificationDrawer";
